requirements.txt
# 抖音解析 Skill - Python 依赖
requests>=2.31.0
python-dotenv>=1.0.0
gmssl>=3.2.0
